extract AzureMessage
